+++ packages/patternlab/styleguide/source/assets/js/index.js
@@ -5,8 +5,6 @@ import ajaxPattern from "./modules/ajaxPattern.js";
 import ajaxPatternOverride from "./modules/ajaxPatternOverride.js";
 import accordions from "./modules/accordions.js";
 import activateBodyTextTruncation from "./modules/activateBodyTextTruncation.js";
-import googleMap from "./modules/googleMap.js";
-import leafletMap from "./modules/leafletMap.js";
 import back2top from "./modules/back2top.js";
 import brandBanner from "./modules/brandBanner.js";
 import clickable from "./modules/clickable.js";
@@ -32,6 +30,8 @@ import locationFilters from "./modules/locationFilters.js";
 import mainNav from "./modules/mainNav.js";
 import mainNavHamburger from "./modules/mainNavHamburger.js";
 import mainNavMIxed from "./modules/mainNavMixed.js";
+import mapGoogle from "./modules/mapGoogle.js";
+import mapLeaflet from "./modules/mapLeaflet.js";
 import mobileNav from "./modules/mobileNav.js";
 import orgSelector from "./modules/orgSelector.js";
 import organizationNavigation from "./modules/organizationNavigation.js";
