+++ packages/patternlab/styleguide/source/assets/js/modules/mapGoogle.js
